Fix @ignore at multiple example sets (#1679)
* first splitting of UnitTestFeatureGenerator

* reformat code

* start adding if to the code

* adjust test generation for ignore tags in example sets

* fix vb.net code generation

* small refactorings
