+++ TechTalk.SpecFlow.Generator/Generation/GeneratorConstants.cs
@@ -1,0 +1,19 @@
+﻿namespace TechTalk.SpecFlow.Generator.Generation
+{
+    public class GeneratorConstants
+    {
+        public const string DEFAULT_NAMESPACE = "SpecFlowTests";
+        public const string TEST_NAME_FORMAT = "{0}";
+        public const string SCENARIO_INITIALIZE_NAME = "ScenarioInitialize";
+        public const string SCENARIO_START_NAME = "ScenarioStart";
+        public const string SCENARIO_CLEANUP_NAME = "ScenarioCleanup";
+        public const string TEST_INITIALIZE_NAME = "TestInitialize";
+        public const string TEST_CLEANUP_NAME = "TestTearDown";
+        public const string TESTCLASS_INITIALIZE_NAME = "FeatureSetup";
+        public const string TESTCLASS_CLEANUP_NAME = "FeatureTearDown";
+        public const string BACKGROUND_NAME = "FeatureBackground";
+        public const string TESTRUNNER_FIELD = "testRunner";
+        public const string SPECFLOW_NAMESPACE = "TechTalk.SpecFlow";
+        public const string SCENARIO_OUTLINE_EXAMPLE_TAGS_PARAMETER = "exampleTags";
+    }
+}
