+++ Tests/TechTalk.SpecFlow.GeneratorTests/FeatureGeneratorProviderTests.cs
@@ -4,6 +4,7 @@ using TechTalk.SpecFlow.Generator.UnitTestConverter;
 using FluentAssertions;
 using TechTalk.SpecFlow.Configuration;
 using TechTalk.SpecFlow.Generator.CodeDom;
+using TechTalk.SpecFlow.Generator.Generation;
 using TechTalk.SpecFlow.Generator.UnitTestProvider;
 
 namespace TechTalk.SpecFlow.GeneratorTests
