+++ Tests/TechTalk.SpecFlow.GeneratorTests/FeatureGeneratorRegistryTests.cs
@@ -7,6 +7,7 @@ using Moq;
 using Xunit;
 using TechTalk.SpecFlow.Configuration;
 using TechTalk.SpecFlow.Generator;
+using TechTalk.SpecFlow.Generator.Generation;
 using TechTalk.SpecFlow.Generator.Interfaces;
 using TechTalk.SpecFlow.Generator.UnitTestConverter;
 using TechTalk.SpecFlow.Parser;
