+++ Tests/TechTalk.SpecFlow.GeneratorTests/IUnitTestGeneratorProviderExtensions.cs
@@ -2,6 +2,7 @@
 using TechTalk.SpecFlow.Generator;
 using TechTalk.SpecFlow.Generator.CodeDom;
 using TechTalk.SpecFlow.Generator.Configuration;
+using TechTalk.SpecFlow.Generator.Generation;
 using TechTalk.SpecFlow.Generator.UnitTestProvider;
 using TechTalk.SpecFlow.Utils;
