+++ Tests/TechTalk.SpecFlow.Specs.Generator.SpecFlowPlugin/CombinationFeatureGenerator.cs
@@ -1,6 +1,7 @@
 ﻿using TechTalk.SpecFlow.Configuration;
 using TechTalk.SpecFlow.Generator;
 using TechTalk.SpecFlow.Generator.CodeDom;
+using TechTalk.SpecFlow.Generator.Generation;
 using TechTalk.SpecFlow.Generator.UnitTestConverter;
 
 namespace TechTalk.SpecFlow.Specs.Generator.SpecFlowPlugin
